+++ portofino-database/src/main/java/com/manydesigns/portofino/database/multitenancy/MultiTenant.java
@@ -1,0 +1,7 @@
+package com.manydesigns.portofino.database.multitenancy;
+
+public @interface MultiTenant {
+
+    Class strategy();
+
+}