+++ portofino-database/src/main/java/com/manydesigns/portofino/modules/DatabaseModule.java
@@ -25,6 +25,7 @@ import com.manydesigns.portofino.code.AggregateCodeBase;
 import com.manydesigns.portofino.code.CodeBase;
 import com.manydesigns.portofino.model.database.platforms.DatabasePlatformsRegistry;
 import com.manydesigns.portofino.persistence.Persistence;
+import com.manydesigns.portofino.persistence.hibernate.multitenancy.MultiTenancyImplementationFactory;
 import com.manydesigns.portofino.spring.PortofinoSpringConfiguration;
 import io.reactivex.disposables.Disposable;
 import org.apache.commons.configuration2.Configuration;
@@ -42,7 +43,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
-import org.springframework.context.ApplicationEvent;
 import org.springframework.context.ApplicationListener;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.event.ContextRefreshedEvent;
@@ -128,11 +128,27 @@ public class DatabaseModule implements Module, ApplicationContextAware, Applicat
     }
 
     @Bean
+    public MultiTenancyImplementationFactory getApplicationContextMTImplFactory() {
+        return implClass -> {
+            try {
+                return applicationContext.getBean(implClass);
+            } catch (BeansException e) {
+                logger.error("MultiTenancyImplementation is not a valid spring bean, trying default constructor");
+                return MultiTenancyImplementationFactory.DEFAULT.make(implClass);
+            }
+        };
+    }
+
+    @Bean
     public Persistence getPersistence(
             @Autowired DatabasePlatformsRegistry databasePlatformsRegistry,
             @Autowired CacheResetListenerRegistry cacheResetListenerRegistry) throws FileSystemException {
-        Persistence persistence = new Persistence(applicationDirectory, configuration, configurationFile, databasePlatformsRegistry);
+        Persistence persistence = new Persistence(
+                applicationDirectory, configuration, configurationFile, databasePlatformsRegistry);
         persistence.cacheResetListenerRegistry = cacheResetListenerRegistry;
+        if(applicationContext != null) { //We may want it to be null when testing
+            applicationContext.getAutowireCapableBeanFactory().autowireBean(persistence);
+        }
 
         FileObject generatedClassesRoot = applicationDirectory.resolveFile(GENERATED_CLASSES_DIRECTORY_NAME);
         generatedClassesRoot.createFolder();