+++ portofino-model/src/main/java/com/manydesigns/portofino/model/database/Database.java
@@ -76,7 +76,7 @@ public class Database implements ModelObject, Annotated {
     }
 
     //**************************************************************************
-    // DatamodelObject implementation
+    // ModelObject implementation
     //**************************************************************************
 
     public void afterUnmarshal(Unmarshaller u, Object parent) {}
@@ -171,10 +171,6 @@ public class Database implements ModelObject, Annotated {
         }
         return result;
     }
-
-    //**************************************************************************
-    // toString() override
-    //**************************************************************************
 
     @Override
     public String toString() {